♻️ refactor : 텍스트 스타일 코드 분리 #201
